Complete very basic view-wallet functionality.
very basic view-only wallet features work.

Added a jank testbed that will eventually be removed.  This testbed is used to
confirm the very primitive view wallet functionality.  It is committed mostly
for posterity.

Receives and Spends appear to be working correctly at this point for the main address.
Subaddresses are yet to come.